// lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading Journal API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TradeSide = typeof TradeSide[keyof typeof TradeSide];


export const TradeSide = {
  Long: 'Long',
  Short: 'Short',
} as const;

export type TradeResult = typeof TradeResult[keyof typeof TradeResult];


export const TradeResult = {
  Win: 'Win',
  Loss: 'Loss',
} as const;

export interface Trade {
  id: number;
  date: string;
  ticker: string;
  side: TradeSide;
  /** @nullable */
  strategy?: string | null;
  entryPrice: number;
  exitPrice: number;
  stopLoss: number;
  shares: number;
  /** @nullable */
  notes?: string | null;
  netPnl: number;
  rr: number;
  result: TradeResult;
  createdAt: string;
}

export type TradeInputSide = typeof TradeInputSide[keyof typeof TradeInputSide];


export const TradeInputSide = {
  Long: 'Long',
  Short: 'Short',
} as const;

export interface TradeInput {
  date: string;
  ticker: string;
  side: TradeInputSide;
  strategy?: string;
  entryPrice: number;
  exitPrice: number;
  stopLoss: number;
  shares: number;
  notes?: string;
}

export type PnlPointResult = typeof PnlPointResult[keyof typeof PnlPointResult];


export const PnlPointResult = {
  Win: 'Win',
  Loss: 'Loss',
} as const;

export interface PnlPoint {
  tradeNumber: number;
  pnl: number;
  result: PnlPointResult;
  ticker: string;
  date: string;
}

export interface TradeStats {
  totalTrades: number;
  winRate: number;
  netPnl: number;
  profitFactor: number;
  wins: number;
  losses: number;
  pnlByTrade: PnlPoint[];
}

export type ListTradesParams = {
search?: string;
side?: ListTradesSide;
result?: ListTradesResult;
sortOrder?: ListTradesSortOrder;
};

export type ListTradesSide = typeof ListTradesSide[keyof typeof ListTradesSide];


export const ListTradesSide = {
  Long: 'Long',
  Short: 'Short',
} as const;

export type ListTradesResult = typeof ListTradesResult[keyof typeof ListTradesResult];


export const ListTradesResult = {
  Win: 'Win',
  Loss: 'Loss',
} as const;

export type ListTradesSortOrder = typeof ListTradesSortOrder[keyof typeof ListTradesSortOrder];


export const ListTradesSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type DeleteTrade200 = {
  success: boolean;
};